Sorting feat: Enabled user sorting columns by header names

+++ src/components/catalogue/catalogue.tsx
@@ -1,17 +1,39 @@
-import { useState } from "react";
-import Alert from "@mui/material/Alert";
+import { useEffect, useState } from "react";
 import type { MusicEntry } from "../../utils/types";
+import Alert from "@mui/material/Alert";
 import PrimaryButton from "../primary-button";
 import { AddIcon, CheckIcon, ErrorOutlineIcon } from "../utils/mui-icons";
 import { DeleteBtn } from "./delete-button";
 import { deleteAlbumEntry } from "../../utils/api/delete-album-entry";
+import { SortButton } from "./sort-icon";
+import { sortEvents } from "../../utils/events/sort-events";
 
 export const Catalogue = ({ albums }: { albums: MusicEntry[] }) => {
   const PAGE_COUNT = 20;
   const [currentPage, setCurrentPage] = useState(1);
+  const [sortedAlbums, setSortedAlbums] = useState<MusicEntry[]>(albums);
+  const [sortKey, setSortKey] = useState<string>("");
+  const [sortDirection, setSortDirection] = useState<"asc" | "desc">("asc");
   const [isDeleting, setIsDeleting] = useState(false);
   const [userMessage, setUserMessage] = useState("");
   const [error, setError] = useState(false);
+
+  useEffect(() => {
+    if (sortKey) {
+      setSortedAlbums(sortEvents(sortKey, albums, sortDirection));
+    } else {
+      setSortedAlbums(albums);
+    }
+  }, [albums, sortKey, sortDirection]);
+
+  const handleSort = (key: string) => {
+    const sameColumn = sortKey === key;
+    const newDirection = sameColumn && sortDirection === "asc" ? "desc" : "asc";
+
+    setSortKey(key);
+    setSortDirection(newDirection);
+    setCurrentPage(1);
+  };
 
   const handleSeeMore = () => {
     setCurrentPage(currentPage + 1);
@@ -19,6 +41,12 @@ export const Catalogue = ({ albums }: { albums: MusicEntry[] }) => {
 
   const handleDelete = async (id: number) => {
     if (isDeleting) return;
+
+    const confirm = window.confirm(
+      "Are you sure you want to delete this album?",
+    );
+    if (!confirm) return;
+
     setIsDeleting(true);
     try {
       await deleteAlbumEntry(id);
@@ -38,7 +66,7 @@ export const Catalogue = ({ albums }: { albums: MusicEntry[] }) => {
   return (
     <div>
       <div>
-        <h1>Collection ({albums.length})</h1>
+        <h1>Collection ({sortedAlbums.length})</h1>
         <p>
           Updated list of albums per artist in my collection. The list is sorted
           by release year and alphabetically by artist name.
@@ -81,14 +109,34 @@ export const Catalogue = ({ albums }: { albums: MusicEntry[] }) => {
         <table>
           <thead>
             <tr>
-              <th>Artist</th>
-              <th>Album</th>
-              <th>Genre</th>
-              <th>Release Year</th>
+              <th>
+                <div className="th-content">
+                  <span>Artist</span>
+                  <SortButton onClick={() => handleSort("artist")} />
+                </div>
+              </th>
+              <th>
+                <div className="th-content">
+                  <span>Album</span>
+                  <SortButton onClick={() => handleSort("album")} />
+                </div>
+              </th>
+              <th>
+                <div className="th-content">
+                  <span>Genre</span>
+                  <SortButton onClick={() => handleSort("genre")} />
+                </div>
+              </th>
+              <th>
+                <div className="th-content">
+                  <span>Release Year</span>
+                  <SortButton onClick={() => handleSort("releaseYear")} />
+                </div>
+              </th>
             </tr>
           </thead>
           <tbody>
-            {albums
+            {sortedAlbums
               .slice(0, currentPage * PAGE_COUNT)
               .map((item: MusicEntry, index: number) => (
                 <tr key={index}>
@@ -113,7 +161,7 @@ export const Catalogue = ({ albums }: { albums: MusicEntry[] }) => {
           </tbody>
         </table>
       </div>
-      {currentPage * PAGE_COUNT < albums.length && (
+      {currentPage * PAGE_COUNT < sortedAlbums.length && (
         <PrimaryButton
           text="See more"
           icon={<AddIcon />}
